Split quantize entry points and model-hygiene passes out of onnxsim.cpp
Continues the onnxsim.cpp breakup (constant_folding.cpp/partial_shape_eval.cpp)
with two more self-contained modules:

- quantize_entry.cpp/h: the single-call quantization entry points exposed to
  Python (QuantizeDynamic/Ternary/WeightOnly[Int4]/Static/QOperator/Fp16/
  Bf16/Fp8 and their List*QuantizableActivations/Outputs helpers), each a
  thin wrapper around one onnxoptimizer pass run standalone via
  OptimizeFixed.
- model_prep.cpp/h: model-hygiene utilities shared by Simplify()/
  SimplifyPath() -- model fingerprinting, the generic FixedPointFn
  combinator that drives every fixed-point loop in the pipeline, custom/
  default-domain schema registration, node naming, incomplete value_info
  cleanup, opset conversion, CSE hashability, and the input-shape/output
  overwrite helpers.

onnxsim.cpp now holds just the top-level orchestration (Simplify/
SimplifyPath, OptAndShapeOnGraph, the single-step debug entry points) and
shrinks from 1749 to 846 lines (~4370 lines originally, before the whole
split). No behavior change; verified by building the onnxsim library, the
ORT-off wheel path (setup.py's actual build) including cpp2py_export.cc,
and running onnxsim.simplify() end to end -- constant folding, the
partial-shape-eval Reshape rewrite, RNN zero-state elimination, the
quantize_dynamic/quantize_fp16 entry points, and the
overwrite_input_shapes/unused_output/node-naming options all produce
correct output.
