+++ onnxsim/onnxsim.cpp
@@ -1,24 +1,16 @@
 #include "onnxsim.h"
 
-#include <google/protobuf/arena.h>
 #include <google/protobuf/io/zero_copy_stream.h>
-#include <google/protobuf/text_format.h>
-#include <google/protobuf/util/message_differencer.h>
 #include <onnx/onnx_pb.h>
 
 #include <algorithm>
 #include <chrono>
 #include <cstdint>
-#include <cstring>
-#include <fstream>
 #include <functional>
 #include <iomanip>
 #include <map>
-#include <mutex>
-#include <numeric>
 #include <set>
 #include <string>
-#include <type_traits>
 #include <unordered_map>
 #include <unordered_set>
 #include <vector>
@@ -36,675 +28,20 @@
 #include "constant_folding.h"
 #include "contrib_schemas.h"
 #include "custom_optimizer_passes.h"
-#include "dlpack_bridge.h"
+#include "model_prep.h"
 #include "onnx/common/file_utils.h"
 #include "onnx/common/graph_shape_inference.h"
 #include "onnx/common/ir_pb_converter.h"
-#include "onnx/common/ir_pb_converter_internal.h"
 #include "onnx/defs/printer.h"
-#include "onnx/defs/schema.h"
 #include "onnx/inliner/inliner.h"
 #include "onnx/shape_inference/implementation.h"
-#include "onnx/version_converter/convert.h"
 #include "onnxoptimizer/model_util.h"
 #include "onnxoptimizer/optimize.h"
 #include "onnxoptimizer/passes/cse_util.h"
 #include "onnxoptimizer/passes/logging.h"
 #include "partial_shape_eval.h"
-#include "passes/quantize_bf16.h"
-#include "passes/quantize_conv_common.h"
-#include "passes/quantize_fp16.h"
-#include "passes/quantize_fp8.h"
-#include "passes/quantize_matmul_common.h"
-#include "passes/static_quantize_matmul.h"
 #include "profiler.h"
-#include "sym_shape_infer.h"
-#include "sym_value_eval.h"
-
-// A 128-bit fingerprint of a model, used by FixedPointFn to detect when an
-// iteration stopped changing the model without keeping a second full ModelProto
-// around just for the comparison. Two models with the same fingerprint are
-// treated as equal; the odds of a false match are ~2^-128 per comparison, and a
-// false match would only stop simplification one round early (the model stays
-// valid), never produce an incorrect model.
-struct ModelFingerprint {
-  uint64_t h1;
-  uint64_t h2;
-  bool operator==(const ModelFingerprint& other) const {
-    return h1 == other.h1 && h2 == other.h2;
-  }
-};
-
-// Mixes ``n`` bytes at ``data`` into the two rolling hash accumulators (FNV-1a
-// and a splitmix-style mix), 8 bytes (one machine word) at a time rather than
-// one byte at a time -- cutting the number of loop iterations, and with them
-// most of the loop/branch overhead, by roughly 8x versus a byte-at-a-time
-// mix, for identical byte coverage and order.
-void MixBytes(const char* data, size_t n, uint64_t& h1, uint64_t& h2) {
-  const size_t n_words = n / sizeof(uint64_t);
-  for (size_t i = 0; i < n_words; ++i) {
-    // memcpy, not a reinterpret_cast dereference: ``data`` is not guaranteed
-    // 8-byte aligned (it may be a chunk of a streaming buffer, or a
-    // std::string's internal buffer), and an unaligned uint64_t load is
-    // undefined behavior in C++. Any decent compiler lowers this fixed-size
-    // memcpy to a single unaligned load.
-    uint64_t word;
-    std::memcpy(&word, data + i * sizeof(uint64_t), sizeof(uint64_t));
-    h1 = (h1 ^ word) * 1099511628211ULL;
-    h2 = (h2 + word) * 0x9E3779B97F4A7C15ULL;
-    h2 ^= h2 >> 29;
-  }
-  // Fewer than 8 trailing bytes: fall back to a byte-at-a-time mix so every
-  // byte still participates in the hash.
-  for (size_t i = n_words * sizeof(uint64_t); i < n; ++i) {
-    const unsigned char c = static_cast<unsigned char>(data[i]);
-    h1 = (h1 ^ c) * 1099511628211ULL;
-    h2 = (h2 + c) * 0x9E3779B97F4A7C15ULL;
-    h2 ^= h2 >> 29;
-  }
-}
-
-// A protobuf output stream that hashes bytes as they are handed out by the
-// serializer instead of collecting them into a buffer. Fingerprint() used to
-// call ``model.SerializeAsString()`` -- which allocates a buffer the size of
-// the whole serialized model (hundreds of MB of initializer weight data on a
-// large model) and serializes into it -- and then made a second, separate
-// pass over that buffer to hash it. Handing the serializer a small, reused
-// scratch buffer instead avoids that large allocation (and the deallocation
-// moments later): each time the serializer fills it, this stream
-// mixes it into the running hash and hands the same buffer back out, so the
-// whole model is hashed in bounded extra memory regardless of its size.
-class HashingOutputStream final
-    : public google::protobuf::io::ZeroCopyOutputStream {
- public:
-  bool Next(void** data, int* size) override {
-    FlushPending();
-    *data = buf_.data();
-    *size = static_cast<int>(buf_.size());
-    pending_ = static_cast<int>(buf_.size());
-    return true;
-  }
-  void BackUp(int count) override { pending_ -= count; }
-  int64_t ByteCount() const override { return byte_count_; }
-
-  // Call after serialization completes to mix in the final partial buffer.
-  void Finish() { FlushPending(); }
-
-  uint64_t h1() const { return h1_; }
-  uint64_t h2() const { return h2_; }
-  size_t total_bytes() const { return static_cast<size_t>(byte_count_); }
-
- private:
-  void FlushPending() {
-    if (pending_ <= 0) {
-      return;
-    }
-    MixBytes(buf_.data(), static_cast<size_t>(pending_), h1_, h2_);
-    byte_count_ += pending_;
-    pending_ = 0;
-  }
-
-  // 64 KiB: large enough that the per-Next() call overhead is negligible next
-  // to the memory-bandwidth-bound hashing work, small enough to stay resident
-  // in L1/L2 cache across the mix. Heap-allocated, not a fixed-size array
-  // member: this object lives on the call stack of Fingerprint(), which is
-  // itself invoked from deep inside the (possibly nested) fixed-point
-  // machinery, and the WASM build's stack is only tens of KiB total -- a
-  // 64 KiB stack-resident array here reliably overflowed it (issue caught by
-  // the WASM build's own smoke test, "Aborted(stack overflow ...)" on the
-  // very first Simplify() call). std::vector's control block is a few
-  // pointer-sized words on the stack; the buffer itself is heap-backed.
-  std::vector<char> buf_ = std::vector<char>(1 << 16);
-  int pending_ = 0;
-  int64_t byte_count_ = 0;
-  uint64_t h1_ = 1469598103934665603ULL;  // FNV-1a offset basis
-  uint64_t h2_ = 0;
-};
-
-ModelFingerprint Fingerprint(const onnx::ModelProto& model) {
-  // ModelProto contains no protobuf ``map<>`` fields, so serialization order is
-  // stable and equal models serialize to identical bytes.
-  //
-  // Fingerprint() runs at every step of every (possibly nested) fixed point,
-  // so on a model with hundreds of MB of initializer weight data -- which
-  // dominates the serialized size but is untouched by most rounds (shape
-  // inference, the onnx-optimizer passes) -- it still has to read all of that
-  // data, every round, just to prove nothing changed. Profiling a 265 MB
-  // model showed this was a multi-second fraction (over half of total wall
-  // time on that model) of total simplification time, so it is worth
-  // streaming through a small buffer (see HashingOutputStream) rather than
-  // materializing the whole serialized model first.
-  HashingOutputStream stream;
-  // SerializeToZeroCopyStream only returns false if a Next() call on the
-  // stream fails (out of disk space, a broken pipe, ...); HashingOutputStream
-  // writes to memory and its Next() always succeeds, so this cannot fail.
-  model.SerializeToZeroCopyStream(&stream);
-  stream.Finish();
-  uint64_t h1 = stream.h1();
-  uint64_t h2 = stream.h2();
-  h2 ^= stream.total_bytes();
-  return {h1, h2};
-}
-
-// Alternately apply ``f1`` and ``f2`` until the model stops changing (a joint
-// fixed point) or ``max_iters`` alternations elapse. Each application produces
-// a fresh model, so ``model`` is move-assigned in place and only a single
-// ModelProto is held live across the loop; convergence is detected by comparing
-// the fingerprints of consecutive states rather than keeping the previous
-// ModelProto for a ``MessageDifferencer::Equals`` call. This mirrors the
-// original consecutive-pair comparison exactly -- it stops as soon as the last
-// applied function left the model unchanged -- while roughly halving the number
-// of full model copies held at once (which matters because these fixed points
-// nest).
-// The transforms mutate the model in place (``std::function<void(T&)>``), so a
-// transform that already works in place (e.g. ``_InferShapes``) makes no copy
-// at all, and one that must build a fresh model (e.g. ``Optimize``, whose
-// underlying ``OptimizeFixed`` returns a new proto) move-assigns it back. The
-// returned function likewise mutates in place, so it composes when these fixed
-// points nest and a single ModelProto is threaded through the whole thing.
-template <typename T>
-std::function<void(T&)> FixedPointFn(const std::function<void(T&)>& f1,
-                                     const std::function<void(T&)>& f2,
-                                     size_t max_iters, bool* converged) {
-  return [f1, f2, max_iters, converged](T& model) -> void {
-    // Profiled separately from the transforms it gates: on large models this
-    // convergence check is not free (see Fingerprint()'s comment), so its cost
-    // should be visible in its own right rather than silently inflating
-    // whichever of f1/f2's spans happens to run next. A no-op unless
-    // ONNXSIM_PROFILE is set.
-    auto fingerprint = [](const onnx::ModelProto& m) {
-      onnxsim::ProfiledScope scope("Fingerprint");
-      return Fingerprint(m);
-    };
-    size_t _max_iters = max_iters;
-    f1(model);
-    ModelFingerprint fp_prev = fingerprint(model);
-    f2(model);
-    ModelFingerprint fp_cur = fingerprint(model);
-    while (_max_iters-- > 0) {
-      if (fp_cur == fp_prev) {
-        if (converged) {
-          *converged = true;
-        }
-        return;
-      }
-      f1(model);
-      fp_prev = fp_cur;
-      fp_cur = fingerprint(model);
-      if (fp_cur == fp_prev) {
-        if (converged) {
-          *converged = true;
-        }
-        return;
-      }
-      f2(model);
-      fp_prev = fp_cur;
-      fp_cur = fingerprint(model);
-    }
-
-    if (converged) {
-      *converged = false;
-    }
-  };
-}
-
-template <typename T>
-std::function<void(T&)> FixedPointFn(const std::function<void(T&)>& f1,
-                                     const std::function<void(T&)>& f2,
-                                     size_t max_iters) {
-  return FixedPointFn(f1, f2, max_iters, nullptr);
-}
-
-// Same convergence algorithm as the ``Fingerprint``-based ``FixedPointFn``
-// above, specialized for transforms that can cheaply report whether they
-// changed the model themselves (``f1``/``f2`` return true iff they did),
-// instead of hashing the whole serialized model after every call. This skips
-// ``Fingerprint()`` entirely, which matters most here because this is the
-// innermost, most-frequently-run fixed point (``OptAndShape`` below runs
-// every simplification round). It is only as safe as ``f1``/``f2``'s own
-// signal: both onnx-optimizer's per-pass transform counts and onnx's
-// InferShapes value-change count are exact for onnxsim's pass list (no
-// pass with an ``Empty``/uncounted analysis type is used), so a ``false``
-// return means that call provably made no change -- not just "probably".
-//
-// Convergence requires a whole *round* (one ``f1`` call followed by one
-// ``f2`` call) to report no change from either, not just the most recent
-// call: ``f1`` reporting false only means it found nothing new, not that
-// ``f2``'s previous call is done making downstream cleanup available (e.g.
-// a Complete-efficiency onnx-optimizer pass firing late in
-// ``OptimizeGraphFixed``'s own pass-list order can leave a now-dead node
-// that an earlier-ordered pass in that same call never got a chance to
-// revisit -- ``f1`` correctly reports no new shape info either way, but
-// ``f2`` still has cleanup work left for its *next* call). Stopping right
-// after a lone ``f1`` false, as an earlier version of this function did,
-// silently left that cleanup undone whenever nothing above this loop
-// happened to re-drive it; matching a whole round instead of a single call
-// closes that gap unconditionally, with no dependence on some caller
-// incidentally running this again.
-template <typename T>
-std::function<void(T&)> FixedPointFn(const std::function<bool(T&)>& f1,
-                                     const std::function<bool(T&)>& f2,
-                                     size_t max_iters, bool* converged) {
-  return [f1, f2, max_iters, converged](T& model) -> void {
-    size_t _max_iters = max_iters;
-    while (_max_iters-- > 0) {
-      const bool c1 = f1(model);
-      const bool c2 = f2(model);
-      if (!c1 && !c2) {
-        if (converged) {
-          *converged = true;
-        }
-        return;
-      }
-    }
-
-    if (converged) {
-      *converged = false;
-    }
-  };
-}
-
-template <typename T>
-std::function<void(T&)> FixedPointFn(const std::function<bool(T&)>& f1,
-                                     const std::function<bool(T&)>& f2,
-                                     size_t max_iters) {
-  return FixedPointFn(f1, f2, max_iters, nullptr);
-}
-
-// A no-op in-place transform (mutates nothing), used when shape inference or
-// constant folding is disabled.
-void Identity(onnx::ModelProto&) {}
-
-// Recursively collect the op types of operators that live in ONNX's *default*
-// domain but have no registered schema. These are custom operators -- most
-// commonly TensorRT plugins such as ``BatchedNMS_TRT`` or ``EfficientNMS_TRT``
-// -- that were exported into the default domain instead of a vendor-specific
-// one.
-//
-// Custom ops that already live in a non-default domain (e.g. ``com.microsoft``
-// or ``TRT``) are intentionally ignored: onnx::checker::check_model already
-// tolerates unknown ops in non-standard domains, which is exactly the manual
-// workaround reported in GitHub issue #220.
-void CollectCustomDefaultDomainOps(const onnx::GraphProto& graph,
-                                   int default_opset_version,
-                                   std::set<std::string>& custom_ops) {
-  for (const auto& node : graph.node()) {
-    const std::string& domain = node.domain();
-    const bool is_default_domain = domain.empty() || domain == "ai.onnx";
-    if (is_default_domain &&
-        onnx::OpSchemaRegistry::Schema(node.op_type(), default_opset_version,
-                                       /*domain=*/"") == nullptr) {
-      custom_ops.insert(node.op_type());
-    }
-    // Recurse into subgraphs held in node attributes (If/Loop/Scan bodies).
-    for (const auto& attr : node.attribute()) {
-      if (attr.has_g()) {
-        CollectCustomDefaultDomainOps(attr.g(), default_opset_version,
-                                      custom_ops);
-      }
-      for (const auto& subgraph : attr.graphs()) {
-        CollectCustomDefaultDomainOps(subgraph, default_opset_version,
-                                      custom_ops);
-      }
-    }
-  }
-}
-
-// Register a permissive placeholder schema for every default-domain custom op
-// found in ``model``. Without a schema, onnx::checker::check_model rejects the
-// model with "No Op registered for <op> with domain_version of <n>" and
-// simplification never even starts (GitHub issues #107 and #220). The
-// placeholder accepts any number of inputs/outputs of any tensor type and any
-// attributes, so the checker passes and the op is preserved untouched through
-// simplification. It carries no shape/type inference function, so shape
-// inference simply flows past the op as before.
-void RegisterCustomDefaultDomainOpSchemas(const onnx::ModelProto& model) {
-  int default_opset_version = 1;
-  for (const auto& opset : model.opset_import()) {
-    if (opset.domain().empty() || opset.domain() == "ai.onnx") {
-      default_opset_version =
-          std::max(default_opset_version, static_cast<int>(opset.version()));
-    }
-  }
-
-  std::set<std::string> custom_ops;
-  CollectCustomDefaultDomainOps(model.graph(), default_opset_version,
-                                custom_ops);
-
-  for (const auto& op_type : custom_ops) {
-    onnx::OpSchema schema;
-    schema.SetName(op_type)
-        .SetDomain("")
-        .SinceVersion(1)
-        .SetDoc(
-            "Placeholder schema registered by onnxsim for a custom operator "
-            "(e.g. a TensorRT plugin) exported into the default ONNX domain, "
-            "so "
-            "that the model passes validation and is simplified with the "
-            "operator preserved unchanged.")
-        .Input(0, "inputs", "Variadic inputs of the custom operator.", "T",
-               onnx::OpSchema::Variadic, /*is_homogeneous=*/false,
-               /*min_arity=*/0)
-        .Output(0, "outputs", "Variadic outputs of the custom operator.", "T",
-                onnx::OpSchema::Variadic, /*is_homogeneous=*/false,
-                /*min_arity=*/0)
-        .TypeConstraint("T", onnx::OpSchema::all_tensor_types(),
-                        "Allow inputs and outputs of any tensor type.")
-        // Custom ops carry arbitrary, plugin-specific attributes; accept them
-        // all rather than trying to enumerate them.
-        .AllowUncheckedAttributes();
-    // Never fail or throw: a duplicate registration (e.g. simplifying two
-    // models that use the same custom op in one process) is a harmless no-op.
-    onnx::RegisterSchema(std::move(schema), /*opset_version_to_load=*/1,
-                         /*fail_duplicate_schema=*/false,
-                         /*fail_with_exception=*/false);
-  }
-}
-
-// Collect the names of every node in `graph`, descending into subgraphs held in
-// node attributes (If/Loop/Scan bodies). Used to de-duplicate the names later
-// assigned to nameless nodes.
-void CollectNodeNames(const onnx::GraphProto& graph,
-                      std::set<std::string>& names) {
-  for (const auto& node : graph.node()) {
-    if (!node.name().empty()) {
-      names.insert(node.name());
-    }
-    for (const auto& attr : node.attribute()) {
-      if (attr.has_g()) {
-        CollectNodeNames(attr.g(), names);
-      }
-      for (const auto& subgraph : attr.graphs()) {
-        CollectNodeNames(subgraph, names);
-      }
-    }
-  }
-}
-
-// Give a unique, deterministic name to every node that has none. Nodes without
-// a name survive simplification unnamed -- either because they were nameless in
-// the input model or because an onnx-optimizer pass created a replacement node
-// without setting a name -- which trips up downstream tooling that keys on node
-// names (issue #269). Each generated name is derived from the op type plus a
-// running counter and de-duplicated against every name already present in the
-// graph (including names generated earlier in this pass). Subgraphs are handled
-// recursively so nodes inside If/Loop/Scan bodies are named too.
-void AssignMissingNodeNames(onnx::GraphProto& graph,
-                            std::set<std::string>& used_names,
-                            size_t& counter) {
-  for (auto& node : *graph.mutable_node()) {
-    if (node.name().empty()) {
-      std::string name;
-      do {
-        name = node.op_type() + "_" + std::to_string(counter++);
-      } while (used_names.count(name) > 0);
-      used_names.insert(name);
-      node.set_name(name);
-    }
-    for (auto& attr : *node.mutable_attribute()) {
-      if (attr.has_g()) {
-        AssignMissingNodeNames(*attr.mutable_g(), used_names, counter);
-      }
-      for (auto& subgraph : *attr.mutable_graphs()) {
-        AssignMissingNodeNames(subgraph, used_names, counter);
-      }
-    }
-  }
-}
-
-// Assign names to any nodes left nameless after simplification (issue #269).
-void AssignMissingNodeNames(onnx::ModelProto& model) {
-  std::set<std::string> used_names;
-  CollectNodeNames(model.graph(), used_names);
-  size_t counter = 0;
-  AssignMissingNodeNames(*model.mutable_graph(), used_names, counter);
-}
-
-// Shape inference can leave behind a value_info entry that records a shape
-// but never resolved an element type (e.g. observed on a real-world model's
-// Reshape outputs inside attention blocks: graph_shape_inference.h's
-// InferShapesOnGraph propagated the target shape but left elem_type at its
-// default UNDEFINED). value_info is purely optional annotation -- nothing
-// reads it as ground truth -- but an entry with elem_type == UNDEFINED is a
-// malformed TypeProto, and onnx::checker::check_model does not reject it,
-// while onnxruntime's own model loader is stricter and refuses to load the
-// model at all with "Invalid tensor data type 0". Drop such entries instead
-// of leaving broken metadata in the output model; the same op's actual
-// output tensor is unaffected either way.
-// Recurses into subgraphs (If/Loop/Scan bodies) for the same reason
-// AssignMissingNodeNames does.
-void DropIncompleteValueInfo(onnx::GraphProto& graph) {
-  auto& value_info = *graph.mutable_value_info();
-  value_info.erase(
-      std::remove_if(value_info.begin(), value_info.end(),
-                     [](const onnx::ValueInfoProto& vi) {
-                       return vi.type().has_tensor_type() &&
-                              vi.type().tensor_type().elem_type() ==
-                                  onnx::TensorProto::UNDEFINED;
-                     }),
-      value_info.end());
-  for (auto& node : *graph.mutable_node()) {
-    for (auto& attr : *node.mutable_attribute()) {
-      if (attr.has_g()) {
-        DropIncompleteValueInfo(*attr.mutable_g());
-      }
-      for (auto& subgraph : *attr.mutable_graphs()) {
-        DropIncompleteValueInfo(subgraph);
-      }
-    }
-  }
-}
-
-void DropIncompleteValueInfo(onnx::ModelProto& model) {
-  DropIncompleteValueInfo(*model.mutable_graph());
-}
-
-void Check(const onnx::ModelProto& model) { onnx::checker::check_model(model); }
-
-// Return the opset version the model imports for the default ONNX domain
-// (represented as either the empty string or "ai.onnx"), or std::nullopt when
-// the model does not import the default domain at all (a model made purely of
-// custom-domain operators).
-std::optional<int> DefaultOpsetVersion(const onnx::ModelProto& model) {
-  for (const auto& opset : model.opset_import()) {
-    if (opset.domain().empty() || opset.domain() == "ai.onnx") {
-      return static_cast<int>(opset.version());
-    }
-  }
-  return std::nullopt;
-}
-
-// Mirrors onnx_simplifier.py's remove_initializer_from_input: an initializer
-// that also appears in graph.input is treated as a runtime input by
-// onnxoptimizer's is_constant_initializer, which blocks value-baking fusions
-// (fuse_bn_into_conv, ...) that only work on non-input initializers -- e.g.
-// the plain Conv+BN chains of the opset-8 resnet101-v1-7 were left completely
-// unsimplified without this. Removing it from the input list lets it fold
-// like any other constant. Kept in sync with the Python function; see that
-// one's own comment for the opset-6 floor (Cast's `to` attribute's INT
-// encoding wasn't valid before opset 6, so bumping an old-IR/old-opset model
-// to IR 4 here would let a later fusion emit a node the opset rejects).
-void RemoveInitializerFromInput(onnx::ModelProto& model) {
-  constexpr int kMinOpsetForInitializerFold = 6;
-  if (model.ir_version() < 4) {
-    const auto opset = DefaultOpsetVersion(model);
-    if (!opset || *opset < kMinOpsetForInitializerFold) {
-      return;
-    }
-  }
-  std::set<std::string> initializer_names;
-  for (const auto& init : model.graph().initializer()) {
-    initializer_names.insert(init.name());
-  }
-  auto* inputs = model.mutable_graph()->mutable_input();
-  bool removed_any = false;
-  for (int i = inputs->size() - 1; i >= 0; --i) {
-    if (initializer_names.count((*inputs)[i].name()) > 0) {
-      inputs->erase(inputs->begin() + i);
-      removed_any = true;
-    }
-  }
-  if (removed_any && model.ir_version() < 4) {
-    model.set_ir_version(4);
-  }
-}
-
-// ONNX TensorProto element types onnxoptimizer's tensor-value hashing
-// (cse_util.h) knows how to hash. Any other type makes
-// eliminate_common_subexpression/eliminate_duplicate_initializer raise
-// "no supported data type: <N>". Enumerates the *supported* types (rather
-// than the unsupported ones) so an element type added to ONNX in the future
-// is treated as unhashable by default instead of silently crashing. Mirrors
-// onnx_simplifier.py's _CSE_HASHABLE_ELEM_TYPES.
-bool IsCSEHashableElemType(int32_t elem_type) {
-  switch (elem_type) {
-    case onnx::TensorProto::UNDEFINED:
-    case onnx::TensorProto::BOOL:
-    case onnx::TensorProto::INT8:
-    case onnx::TensorProto::INT16:
-    case onnx::TensorProto::INT32:
-    case onnx::TensorProto::INT64:
-    case onnx::TensorProto::UINT8:
-    case onnx::TensorProto::UINT16:
-    case onnx::TensorProto::UINT32:
-    case onnx::TensorProto::UINT64:
-    case onnx::TensorProto::FLOAT:
-    case onnx::TensorProto::DOUBLE:
-    case onnx::TensorProto::FLOAT16:
-    case onnx::TensorProto::BFLOAT16:
-    case onnx::TensorProto::COMPLEX64:
-    case onnx::TensorProto::COMPLEX128:
-    case onnx::TensorProto::STRING:
-      return true;
-    default:
-      return false;
-  }
-}
-
-// Mirrors onnx_simplifier.py's _has_cse_unhashable_tensor /
-// _iter_tensor_data_types: walks every tensor CSE might hash -- initializers,
-// t/ts node attributes, recursing into subgraphs -- looking for an element
-// type IsCSEHashableElemType rejects.
-bool GraphHasCSEUnhashableTensor(const onnx::GraphProto& graph) {
-  for (const auto& init : graph.initializer()) {
-    if (!IsCSEHashableElemType(init.data_type())) {
-      return true;
-    }
-  }
-  for (const auto& node : graph.node()) {
-    for (const auto& attr : node.attribute()) {
-      if (attr.has_t() && !IsCSEHashableElemType(attr.t().data_type())) {
-        return true;
-      }
-      for (const auto& t : attr.tensors()) {
-        if (!IsCSEHashableElemType(t.data_type())) {
-          return true;
-        }
-      }
-      if (attr.has_g() && GraphHasCSEUnhashableTensor(attr.g())) {
-        return true;
-      }
-      for (const auto& g : attr.graphs()) {
-        if (GraphHasCSEUnhashableTensor(g)) {
-          return true;
-        }
-      }
-    }
-  }
-  return false;
-}
-
-// Mirrors onnx_simplifier.py's overwrite_input_shapes loop: for each named
-// graph input, overwrite its shape's dims with the given values, skipping
-// any non-positive entry (which means "keep the original, possibly dynamic,
-// dimension" rather than hardcoding an invalid size such as 0 -- GitHub
-// issue #237). Throws std::runtime_error if a name isn't a graph input
-// (matching onnx_simplifier.py's RuntimeError for the same case).
-void ApplyInputShapeOverwrite(
-    onnx::ModelProto& model,
-    const std::unordered_map<std::string, std::vector<int64_t>>&
-        overwrite_input_shapes) {
-  for (const auto& [name, shape] : overwrite_input_shapes) {
-    bool found = false;
-    for (auto& input : *model.mutable_graph()->mutable_input()) {
-      if (input.name() != name) {
-        continue;
-      }
-      found = true;
-      auto* dims = input.mutable_type()
-                       ->mutable_tensor_type()
-                       ->mutable_shape()
-                       ->mutable_dim();
-      for (int i = 0; i < dims->size() && i < static_cast<int>(shape.size());
-           ++i) {
-        if (shape[i] > 0) {
-          dims->Mutable(i)->set_dim_value(shape[i]);
-        }
-      }
-    }
-    if (!found) {
-      throw std::runtime_error("The model doesn't have input named \"" + name +
-                               "\"");
-    }
-  }
-}
-
-// Mirrors onnx_simplifier.py's remove_unused_output: drops the named graph
-// outputs. Downstream dead-end elimination cleans up nodes that only fed
-// them. Throws std::runtime_error if a name isn't a graph output (matching
-// onnx_simplifier.py's RuntimeError for the same case).
-void RemoveUnusedOutputs(onnx::ModelProto& model,
-                         const std::vector<std::string>& unused_output) {
-  for (const auto& name : unused_output) {
-    bool found = false;
-    for (const auto& output : model.graph().output()) {
-      if (output.name() == name) {
-        found = true;
-        break;
-      }
-    }
-    if (!found) {
-      throw std::runtime_error("The model doesn't have output named \"" + name +
-                               "\"");
-    }
-  }
-  auto* outputs = model.mutable_graph()->mutable_output();
-  google::protobuf::RepeatedPtrField<onnx::ValueInfoProto> kept;
-  for (auto& output : *outputs) {
-    if (std::find(unused_output.begin(), unused_output.end(), output.name()) ==
-        unused_output.end()) {
-      *kept.Add() = std::move(output);
-    }
-  }
-  outputs->Swap(&kept);
-}
-
-// Convert the default ONNX domain of the model to target_version using onnx's
-// own version converter. Only the default ONNX domain opset is changed;
-// custom/other-domain opset imports are left untouched. Returns the model
-// unchanged when it is already at the requested version or does not import the
-// default domain.
-onnx::ModelProto ConvertOpsetVersion(const onnx::ModelProto& model,
-                                     int target_version) {
-  const auto current = DefaultOpsetVersion(model);
-  if (!current || *current == target_version) {
-    return model;
-  }
-  return onnx::version_conversion::ConvertVersion(model, target_version);
-}
-
-// Shared schema setup for the single-pass debug helpers below, mirroring the
-// head of ``Simplify``: teach shape inference about ONNX Runtime's quantized
-// contrib ops and register permissive placeholders for custom ops exported into
-// the default ONNX domain, so neither shape inference nor a later checker
-// rejects the model. Unlike ``Simplify`` these helpers do not ``Check`` the
-// model up front -- the point of running a step in isolation is to debug a
-// model that may not yet be fully valid.
-void PrepareSchemasForDebug(const onnx::ModelProto& model) {
-  onnxsim::RegisterContribOpSchemas();
-  FixupSchemaDeterminism();
-  RegisterCustomDefaultDomainOpSchemas(model);
-}
+#include "quantize_entry.h"
 
 onnx::ModelProto InferShapesOnce(const onnx::ModelProto& model) {
   PrepareSchemasForDebug(model);
@@ -737,246 +74,6 @@ onnx::ModelProto FoldConstantOnce(const ModelExecutor& executor,
   _EvalPartialShape(out);
   out = _FoldConstant(executor, std::move(out));
   return out;
-}
-
-onnx::ModelProto QuantizeDynamic(const onnx::ModelProto& model) {
-  PrepareSchemasForDebug(model);
-  // Registers dynamic_quantize_matmul (idempotent) into onnxoptimizer's
-  // registry so OptimizeFixed can find it by name below.
-  onnxsim::RegisterCustomOptimizerPasses();
-  return onnx::optimization::OptimizeFixed(
-      model, std::vector<std::string>{"dynamic_quantize_matmul"});
-}
-
-onnx::ModelProto QuantizeTernary(const onnx::ModelProto& model) {
-  PrepareSchemasForDebug(model);
-  // Registers dynamic_quantize_ternary_matmul (idempotent) into
-  // onnxoptimizer's registry so OptimizeFixed can find it by name below.
-  onnxsim::RegisterCustomOptimizerPasses();
-  return onnx::optimization::OptimizeFixed(
-      model, std::vector<std::string>{"dynamic_quantize_ternary_matmul"});
-}
-
-onnx::ModelProto QuantizeWeightOnly(const onnx::ModelProto& model) {
-  PrepareSchemasForDebug(model);
-  // Registers weight_only_quantize_matmul/_conv (idempotent) into
-  // onnxoptimizer's registry so OptimizeFixed can find them by name below.
-  onnxsim::RegisterCustomOptimizerPasses();
-  return onnx::optimization::OptimizeFixed(
-      model, std::vector<std::string>{"weight_only_quantize_matmul",
-                                      "weight_only_quantize_conv"});
-}
-
-onnx::ModelProto QuantizeWeightOnlyInt4(const onnx::ModelProto& model) {
-  PrepareSchemasForDebug(model);
-  // Registers weight_only_quantize_int4_matmul/_conv (idempotent) into
-  // onnxoptimizer's registry so OptimizeFixed can find them by name below.
-  onnxsim::RegisterCustomOptimizerPasses();
-  return onnx::optimization::OptimizeFixed(
-      model, std::vector<std::string>{"weight_only_quantize_int4_matmul",
-                                      "weight_only_quantize_int4_conv"});
-}
-
-std::vector<std::string> ListQuantizableActivations(
-    const onnx::ModelProto& model) {
-  PrepareSchemasForDebug(model);
-  std::shared_ptr<onnx::Graph> g(onnx::ImportModelProto(model));
-  if (g.get() == nullptr) {
-    return {};
-  }
-  std::vector<std::string> names;
-  std::unordered_set<std::string> seen;
-  for (auto* node : g->nodes()) {
-    onnx::optimization::onnxsim_passes::MatMulLikeInfo info;
-    if (!onnx::optimization::onnxsim_passes::MatchMatMulLike(node, info)) {
-      continue;
-    }
-    if (info.x->elemType() != onnx::TensorProto_DataType_FLOAT) {
-      continue;
-    }
-    const onnx::Tensor* w_t = onnx::optimization::FetchConstantTensor(info.w);
-    if (w_t == nullptr ||
-        w_t->elem_type() != onnx::TensorProto_DataType_FLOAT ||
-        w_t->sizes().size() != 2) {
-      continue;
-    }
-    if (seen.insert(info.x->uniqueName()).second) {
-      names.push_back(info.x->uniqueName());
-    }
-  }
-  for (auto* node : g->nodes()) {
-    onnx::optimization::onnxsim_passes::ConvInfo info;
-    if (!onnx::optimization::onnxsim_passes::MatchConv(node, info)) {
-      continue;
-    }
-    if (info.x->elemType() != onnx::TensorProto_DataType_FLOAT) {
-      continue;
-    }
-    const onnx::Tensor* w_t = onnx::optimization::FetchConstantTensor(info.w);
-    if (w_t == nullptr ||
-        w_t->elem_type() != onnx::TensorProto_DataType_FLOAT ||
-        w_t->sizes().size() < 3) {
-      continue;
-    }
-    if (seen.insert(info.x->uniqueName()).second) {
-      names.push_back(info.x->uniqueName());
-    }
-  }
-  return names;
-}
-
-onnx::ModelProto QuantizeStatic(
-    const onnx::ModelProto& model,
-    const std::unordered_map<std::string, std::pair<float, float>>&
-        activation_ranges) {
-  PrepareSchemasForDebug(model);
-  // Registers static_quantize_matmul (idempotent) into onnxoptimizer's
-  // registry so OptimizeFixed can find it by name below.
-  onnxsim::RegisterCustomOptimizerPasses();
-  // static_quantize_matmul reads this global, the same way onnxsim's other
-  // passes read `config` -- see StaticQuantizationCalibrationRanges's doc
-  // comment.
-  onnx::optimization::onnxsim_passes::StaticQuantizationCalibrationRanges() =
-      activation_ranges;
-  return onnx::optimization::OptimizeFixed(
-      model, std::vector<std::string>{"static_quantize_matmul",
-                                      "static_quantize_conv"});
-}
-
-std::vector<std::string> ListQOperatorQuantizableOutputs(
-    const onnx::ModelProto& model) {
-  PrepareSchemasForDebug(model);
-  std::shared_ptr<onnx::Graph> g(onnx::ImportModelProto(model));
-  if (g.get() == nullptr) {
-    return {};
-  }
-  std::vector<std::string> names;
-  std::unordered_set<std::string> seen;
-  for (auto* node : g->nodes()) {
-    onnx::optimization::onnxsim_passes::MatMulLikeInfo info;
-    if (!onnx::optimization::onnxsim_passes::MatchMatMulLike(node, info)) {
-      continue;
-    }
-    if (info.x->elemType() != onnx::TensorProto_DataType_FLOAT) {
-      continue;
-    }
-    const onnx::Tensor* w_t = onnx::optimization::FetchConstantTensor(info.w);
-    if (w_t == nullptr ||
-        w_t->elem_type() != onnx::TensorProto_DataType_FLOAT ||
-        w_t->sizes().size() != 2) {
-      continue;
-    }
-    if (seen.insert(node->output()->uniqueName()).second) {
-      names.push_back(node->output()->uniqueName());
-    }
-  }
-  for (auto* node : g->nodes()) {
-    onnx::optimization::onnxsim_passes::ConvInfo info;
-    if (!onnx::optimization::onnxsim_passes::MatchConv(node, info)) {
-      continue;
-    }
-    if (info.x->elemType() != onnx::TensorProto_DataType_FLOAT) {
-      continue;
-    }
-    const onnx::Tensor* w_t = onnx::optimization::FetchConstantTensor(info.w);
-    if (w_t == nullptr ||
-        w_t->elem_type() != onnx::TensorProto_DataType_FLOAT ||
-        w_t->sizes().size() < 3) {
-      continue;
-    }
-    if (info.bias != nullptr) {
-      // qoperator_quantize_conv only rewrites a Conv whose bias (if any) is a
-      // constant float32 [Cout] tensor it can pre-quantize to INT32 -- see
-      // that pass's doc comment. Skip listing this Conv's output otherwise,
-      // since it will never actually be rewritten.
-      const onnx::Tensor* b_t =
-          onnx::optimization::FetchConstantTensor(info.bias);
-      if (b_t == nullptr ||
-          b_t->elem_type() != onnx::TensorProto_DataType_FLOAT ||
-          b_t->sizes().size() != 1) {
-        continue;
-      }
-    }
-    if (seen.insert(node->output()->uniqueName()).second) {
-      names.push_back(node->output()->uniqueName());
-    }
-  }
-  return names;
-}
-
-onnx::ModelProto QuantizeQOperator(
-    const onnx::ModelProto& model,
-    const std::unordered_map<std::string, std::pair<float, float>>&
-        activation_ranges) {
-  PrepareSchemasForDebug(model);
-  // Registers qoperator_quantize_matmul/_conv (idempotent) into
-  // onnxoptimizer's registry so OptimizeFixed can find them by name below.
-  onnxsim::RegisterCustomOptimizerPasses();
-  // qoperator_quantize_matmul/_conv read the same calibration-ranges global
-  // static_quantize_matmul/_conv do (see StaticQuantizationCalibrationRanges's
-  // doc comment) -- both are keyed by tensor name, and QOperator format's
-  // ranges are a superset (activation names plus output names) of QDQ
-  // format's, so sharing the map is safe as long as only one of
-  // QuantizeStatic/QuantizeQOperator runs per call, which OptimizeFixed's
-  // single pass-name list here ensures.
-  onnx::optimization::onnxsim_passes::StaticQuantizationCalibrationRanges() =
-      activation_ranges;
-  return onnx::optimization::OptimizeFixed(
-      model, std::vector<std::string>{"qoperator_quantize_matmul",
-                                      "qoperator_quantize_conv"});
-}
-
-onnx::ModelProto QuantizeFp16(const onnx::ModelProto& model,
-                              bool keep_io_types) {
-  PrepareSchemasForDebug(model);
-  // Registers quantize_fp16 (idempotent) into onnxoptimizer's registry so
-  // OptimizeFixed can find it by name below.
-  onnxsim::RegisterCustomOptimizerPasses();
-  // quantize_fp16 reads this the same way static_quantize_matmul.h's passes
-  // read StaticQuantizationCalibrationRanges() -- OptimizeFixed's pass-name
-  // list has no way to carry a parameter directly.
-  onnx::optimization::onnxsim_passes::QuantizeFp16KeepIoTypes() = keep_io_types;
-  return onnx::optimization::OptimizeFixed(
-      model, std::vector<std::string>{"quantize_fp16"});
-}
-
-onnx::ModelProto QuantizeBf16(const onnx::ModelProto& model,
-                              bool keep_io_types) {
-  PrepareSchemasForDebug(model);
-  // Registers quantize_bf16 (idempotent) into onnxoptimizer's registry so
-  // OptimizeFixed can find it by name below.
-  onnxsim::RegisterCustomOptimizerPasses();
-  // quantize_bf16 reads this the same way quantize_fp16 reads
-  // QuantizeFp16KeepIoTypes() -- OptimizeFixed's pass-name list has no way
-  // to carry a parameter directly.
-  onnx::optimization::onnxsim_passes::QuantizeBf16KeepIoTypes() = keep_io_types;
-  return onnx::optimization::OptimizeFixed(
-      model, std::vector<std::string>{"quantize_bf16"});
-}
-
-onnx::ModelProto QuantizeFp8(const onnx::ModelProto& model,
-                             const std::string& format, bool keep_io_types) {
-  onnx::optimization::onnxsim_passes::Float8Format target_format;
-  if (format == "e4m3") {
-    target_format = onnx::optimization::onnxsim_passes::Float8Format::kE4M3FN;
-  } else if (format == "e5m2") {
-    target_format = onnx::optimization::onnxsim_passes::Float8Format::kE5M2;
-  } else {
-    throw std::invalid_argument(
-        "QuantizeFp8: format must be \"e4m3\" or \"e5m2\", got \"" + format +
-        "\"");
-  }
-  PrepareSchemasForDebug(model);
-  // Registers quantize_fp8 (idempotent) into onnxoptimizer's registry so
-  // OptimizeFixed can find it by name below.
-  onnxsim::RegisterCustomOptimizerPasses();
-  // quantize_fp8 reads these the same way quantize_fp16 reads
-  // QuantizeFp16KeepIoTypes() -- OptimizeFixed's pass-name list has no way
-  // to carry a parameter directly.
-  onnx::optimization::onnxsim_passes::QuantizeFp8TargetFormat() = target_format;
-  onnx::optimization::onnxsim_passes::QuantizeFp8KeepIoTypes() = keep_io_types;
-  return onnx::optimization::OptimizeFixed(
-      model, std::vector<std::string>{"quantize_fp8"});
 }
 
 // Records the options this Simplify() call actually used (skip_optimizers
