+++ onnxsim/quantize_entry.cpp
@@ -1,0 +1,258 @@
+#include "quantize_entry.h"
+
+#include <onnx/onnx_pb.h>
+
+#include <memory>
+#include <stdexcept>
+#include <unordered_set>
+
+#include "custom_optimizer_passes.h"
+#include "model_prep.h"
+#include "onnx/common/ir_pb_converter.h"
+#include "onnxoptimizer/optimize.h"
+#include "passes/quantize_bf16.h"
+#include "passes/quantize_conv_common.h"
+#include "passes/quantize_fp16.h"
+#include "passes/quantize_fp8.h"
+#include "passes/quantize_matmul_common.h"
+#include "passes/static_quantize_matmul.h"
+
+onnx::ModelProto QuantizeDynamic(const onnx::ModelProto& model) {
+  PrepareSchemasForDebug(model);
+  // Registers dynamic_quantize_matmul (idempotent) into onnxoptimizer's
+  // registry so OptimizeFixed can find it by name below.
+  onnxsim::RegisterCustomOptimizerPasses();
+  return onnx::optimization::OptimizeFixed(
+      model, std::vector<std::string>{"dynamic_quantize_matmul"});
+}
+
+onnx::ModelProto QuantizeTernary(const onnx::ModelProto& model) {
+  PrepareSchemasForDebug(model);
+  // Registers dynamic_quantize_ternary_matmul (idempotent) into
+  // onnxoptimizer's registry so OptimizeFixed can find it by name below.
+  onnxsim::RegisterCustomOptimizerPasses();
+  return onnx::optimization::OptimizeFixed(
+      model, std::vector<std::string>{"dynamic_quantize_ternary_matmul"});
+}
+
+onnx::ModelProto QuantizeWeightOnly(const onnx::ModelProto& model) {
+  PrepareSchemasForDebug(model);
+  // Registers weight_only_quantize_matmul/_conv (idempotent) into
+  // onnxoptimizer's registry so OptimizeFixed can find them by name below.
+  onnxsim::RegisterCustomOptimizerPasses();
+  return onnx::optimization::OptimizeFixed(
+      model, std::vector<std::string>{"weight_only_quantize_matmul",
+                                      "weight_only_quantize_conv"});
+}
+
+onnx::ModelProto QuantizeWeightOnlyInt4(const onnx::ModelProto& model) {
+  PrepareSchemasForDebug(model);
+  // Registers weight_only_quantize_int4_matmul/_conv (idempotent) into
+  // onnxoptimizer's registry so OptimizeFixed can find them by name below.
+  onnxsim::RegisterCustomOptimizerPasses();
+  return onnx::optimization::OptimizeFixed(
+      model, std::vector<std::string>{"weight_only_quantize_int4_matmul",
+                                      "weight_only_quantize_int4_conv"});
+}
+
+std::vector<std::string> ListQuantizableActivations(
+    const onnx::ModelProto& model) {
+  PrepareSchemasForDebug(model);
+  std::shared_ptr<onnx::Graph> g(onnx::ImportModelProto(model));
+  if (g.get() == nullptr) {
+    return {};
+  }
+  std::vector<std::string> names;
+  std::unordered_set<std::string> seen;
+  for (auto* node : g->nodes()) {
+    onnx::optimization::onnxsim_passes::MatMulLikeInfo info;
+    if (!onnx::optimization::onnxsim_passes::MatchMatMulLike(node, info)) {
+      continue;
+    }
+    if (info.x->elemType() != onnx::TensorProto_DataType_FLOAT) {
+      continue;
+    }
+    const onnx::Tensor* w_t = onnx::optimization::FetchConstantTensor(info.w);
+    if (w_t == nullptr ||
+        w_t->elem_type() != onnx::TensorProto_DataType_FLOAT ||
+        w_t->sizes().size() != 2) {
+      continue;
+    }
+    if (seen.insert(info.x->uniqueName()).second) {
+      names.push_back(info.x->uniqueName());
+    }
+  }
+  for (auto* node : g->nodes()) {
+    onnx::optimization::onnxsim_passes::ConvInfo info;
+    if (!onnx::optimization::onnxsim_passes::MatchConv(node, info)) {
+      continue;
+    }
+    if (info.x->elemType() != onnx::TensorProto_DataType_FLOAT) {
+      continue;
+    }
+    const onnx::Tensor* w_t = onnx::optimization::FetchConstantTensor(info.w);
+    if (w_t == nullptr ||
+        w_t->elem_type() != onnx::TensorProto_DataType_FLOAT ||
+        w_t->sizes().size() < 3) {
+      continue;
+    }
+    if (seen.insert(info.x->uniqueName()).second) {
+      names.push_back(info.x->uniqueName());
+    }
+  }
+  return names;
+}
+
+onnx::ModelProto QuantizeStatic(
+    const onnx::ModelProto& model,
+    const std::unordered_map<std::string, std::pair<float, float>>&
+        activation_ranges) {
+  PrepareSchemasForDebug(model);
+  // Registers static_quantize_matmul (idempotent) into onnxoptimizer's
+  // registry so OptimizeFixed can find it by name below.
+  onnxsim::RegisterCustomOptimizerPasses();
+  // static_quantize_matmul reads this global, the same way onnxsim's other
+  // passes read `config` -- see StaticQuantizationCalibrationRanges's doc
+  // comment.
+  onnx::optimization::onnxsim_passes::StaticQuantizationCalibrationRanges() =
+      activation_ranges;
+  return onnx::optimization::OptimizeFixed(
+      model, std::vector<std::string>{"static_quantize_matmul",
+                                      "static_quantize_conv"});
+}
+
+std::vector<std::string> ListQOperatorQuantizableOutputs(
+    const onnx::ModelProto& model) {
+  PrepareSchemasForDebug(model);
+  std::shared_ptr<onnx::Graph> g(onnx::ImportModelProto(model));
+  if (g.get() == nullptr) {
+    return {};
+  }
+  std::vector<std::string> names;
+  std::unordered_set<std::string> seen;
+  for (auto* node : g->nodes()) {
+    onnx::optimization::onnxsim_passes::MatMulLikeInfo info;
+    if (!onnx::optimization::onnxsim_passes::MatchMatMulLike(node, info)) {
+      continue;
+    }
+    if (info.x->elemType() != onnx::TensorProto_DataType_FLOAT) {
+      continue;
+    }
+    const onnx::Tensor* w_t = onnx::optimization::FetchConstantTensor(info.w);
+    if (w_t == nullptr ||
+        w_t->elem_type() != onnx::TensorProto_DataType_FLOAT ||
+        w_t->sizes().size() != 2) {
+      continue;
+    }
+    if (seen.insert(node->output()->uniqueName()).second) {
+      names.push_back(node->output()->uniqueName());
+    }
+  }
+  for (auto* node : g->nodes()) {
+    onnx::optimization::onnxsim_passes::ConvInfo info;
+    if (!onnx::optimization::onnxsim_passes::MatchConv(node, info)) {
+      continue;
+    }
+    if (info.x->elemType() != onnx::TensorProto_DataType_FLOAT) {
+      continue;
+    }
+    const onnx::Tensor* w_t = onnx::optimization::FetchConstantTensor(info.w);
+    if (w_t == nullptr ||
+        w_t->elem_type() != onnx::TensorProto_DataType_FLOAT ||
+        w_t->sizes().size() < 3) {
+      continue;
+    }
+    if (info.bias != nullptr) {
+      // qoperator_quantize_conv only rewrites a Conv whose bias (if any) is a
+      // constant float32 [Cout] tensor it can pre-quantize to INT32 -- see
+      // that pass's doc comment. Skip listing this Conv's output otherwise,
+      // since it will never actually be rewritten.
+      const onnx::Tensor* b_t =
+          onnx::optimization::FetchConstantTensor(info.bias);
+      if (b_t == nullptr ||
+          b_t->elem_type() != onnx::TensorProto_DataType_FLOAT ||
+          b_t->sizes().size() != 1) {
+        continue;
+      }
+    }
+    if (seen.insert(node->output()->uniqueName()).second) {
+      names.push_back(node->output()->uniqueName());
+    }
+  }
+  return names;
+}
+
+onnx::ModelProto QuantizeQOperator(
+    const onnx::ModelProto& model,
+    const std::unordered_map<std::string, std::pair<float, float>>&
+        activation_ranges) {
+  PrepareSchemasForDebug(model);
+  // Registers qoperator_quantize_matmul/_conv (idempotent) into
+  // onnxoptimizer's registry so OptimizeFixed can find them by name below.
+  onnxsim::RegisterCustomOptimizerPasses();
+  // qoperator_quantize_matmul/_conv read the same calibration-ranges global
+  // static_quantize_matmul/_conv do (see StaticQuantizationCalibrationRanges's
+  // doc comment) -- both are keyed by tensor name, and QOperator format's
+  // ranges are a superset (activation names plus output names) of QDQ
+  // format's, so sharing the map is safe as long as only one of
+  // QuantizeStatic/QuantizeQOperator runs per call, which OptimizeFixed's
+  // single pass-name list here ensures.
+  onnx::optimization::onnxsim_passes::StaticQuantizationCalibrationRanges() =
+      activation_ranges;
+  return onnx::optimization::OptimizeFixed(
+      model, std::vector<std::string>{"qoperator_quantize_matmul",
+                                      "qoperator_quantize_conv"});
+}
+
+onnx::ModelProto QuantizeFp16(const onnx::ModelProto& model,
+                              bool keep_io_types) {
+  PrepareSchemasForDebug(model);
+  // Registers quantize_fp16 (idempotent) into onnxoptimizer's registry so
+  // OptimizeFixed can find it by name below.
+  onnxsim::RegisterCustomOptimizerPasses();
+  // quantize_fp16 reads this the same way static_quantize_matmul.h's passes
+  // read StaticQuantizationCalibrationRanges() -- OptimizeFixed's pass-name
+  // list has no way to carry a parameter directly.
+  onnx::optimization::onnxsim_passes::QuantizeFp16KeepIoTypes() = keep_io_types;
+  return onnx::optimization::OptimizeFixed(
+      model, std::vector<std::string>{"quantize_fp16"});
+}
+
+onnx::ModelProto QuantizeBf16(const onnx::ModelProto& model,
+                              bool keep_io_types) {
+  PrepareSchemasForDebug(model);
+  // Registers quantize_bf16 (idempotent) into onnxoptimizer's registry so
+  // OptimizeFixed can find it by name below.
+  onnxsim::RegisterCustomOptimizerPasses();
+  // quantize_bf16 reads this the same way quantize_fp16 reads
+  // QuantizeFp16KeepIoTypes() -- OptimizeFixed's pass-name list has no way
+  // to carry a parameter directly.
+  onnx::optimization::onnxsim_passes::QuantizeBf16KeepIoTypes() = keep_io_types;
+  return onnx::optimization::OptimizeFixed(
+      model, std::vector<std::string>{"quantize_bf16"});
+}
+
+onnx::ModelProto QuantizeFp8(const onnx::ModelProto& model,
+                             const std::string& format, bool keep_io_types) {
+  onnx::optimization::onnxsim_passes::Float8Format target_format;
+  if (format == "e4m3") {
+    target_format = onnx::optimization::onnxsim_passes::Float8Format::kE4M3FN;
+  } else if (format == "e5m2") {
+    target_format = onnx::optimization::onnxsim_passes::Float8Format::kE5M2;
+  } else {
+    throw std::invalid_argument(
+        "QuantizeFp8: format must be \"e4m3\" or \"e5m2\", got \"" + format +
+        "\"");
+  }
+  PrepareSchemasForDebug(model);
+  // Registers quantize_fp8 (idempotent) into onnxoptimizer's registry so
+  // OptimizeFixed can find it by name below.
+  onnxsim::RegisterCustomOptimizerPasses();
+  // quantize_fp8 reads these the same way quantize_fp16 reads
+  // QuantizeFp16KeepIoTypes() -- OptimizeFixed's pass-name list has no way
+  // to carry a parameter directly.
+  onnx::optimization::onnxsim_passes::QuantizeFp8TargetFormat() = target_format;
+  onnx::optimization::onnxsim_passes::QuantizeFp8KeepIoTypes() = keep_io_types;
+  return onnx::optimization::OptimizeFixed(
+      model, std::vector<std::string>{"quantize_fp8"});
+}
